Update l4b_context rt_cmd
Update context such that rt_cmd is the last element in the struct
and remove BUF_SIZE.

NOTE: rt_cmd will be initialized during l4b_init(). It might happen
that malloc (or calloc) is used. In this case, the addition of a
global rt_cmd_buff_freed variable might be useful alongside a
cleanup funtion who's purpose is to de-allocate all allocated memory.